server/requirements.txt
fastapi
uvicorn
python-multipart
onnxruntime
torch
torchvision
Pillow
numpy
onnx